projection/highlight: Rename marked_area -> highlight #544
marked_area isn't a very descriptive name, and will be even less so in
future. Highlight is what we actually call it, and doesn't clash with
old function names, so compatibility methods will be easy to spot.
