+++ OZprivate/rawJS/OZTreeModule/src/projection/projection.js
@@ -3,7 +3,7 @@ import _re_calc  from  './re_calc';
 import _get_shapes from './shape_manager';
 import {calc_horizon as _calc_horizon, update_parent_horizon as _update_parent_horizon, set_horizon_calculator} from './horizon_calc/horizon_calc';
 import data_repo from '../factory/data_repo';
-import { marked_area as _marked_area } from './marked_area/marked_area';
+import { highlight_apply as _highlight_apply } from './highlight/highlight';
 
 class Projection {
   pre_calc(node) {
@@ -15,8 +15,8 @@ class Projection {
   update_parent_horizon(nodes) {
     _update_parent_horizon(nodes);
   }
-  marked_area(node, area_map) {
-    _marked_area(node, area_map)
+  highlight_apply(node, area_map) {
+    _highlight_apply(node, area_map)
   }
   re_calc(node, xp, yp, ws) {
     _re_calc(node, xp, yp, ws);
